feat(ui): replace alert page limit alerts
